+++ src/test/bundler_test.ts
@@ -650,14 +650,37 @@ suite('Bundler', () => {
           doc, matchers.styleMatcher, [], dom5.childNodesIncludeTemplate);
       assert.equal(links.length, 0);
       assert.equal(styles.length, 3);
+      assert.match(dom5.getTextContent(styles[0]), /regular-style/);
+      assert.match(dom5.getTextContent(styles[1]), /simple-style/);
+      assert.match(dom5.getTextContent(styles[2]), /import-linked-style/);
+
+      // Verify the inlined url() values in the stylesheet are not rewritten
+      // to use the "imports/" prefix, since the stylesheet is inside a
+      // `<dom-module>` with an assetpath that defines the base url as
+      // "imports/".
       assert.match(
-          dom5.getTextContent(styles[0]), /regular-style/, 'regular-style.css');
+          dom5.getTextContent(styles[1]), /url\("assets\/platform\.png"\)/);
+    });
+
+    test('Inlined styles observe containing dom-module assetpath', async () => {
+      const doc =
+          await bundle('test/html/style-rewriting.html', {inlineCss: true});
+      const style = dom5.query(
+          doc, matchers.styleMatcher, dom5.childNodesIncludeTemplate)!;
+      assert.isNotNull(style);
+      assert.match(dom5.getTextContent(style), /url\("styles\/unicorn.png"\)/);
+    });
+
+    test('Inlining ignores assetpath if rewriteUrlsInTemplates', async () => {
+      const doc = await bundle(
+          'test/html/style-rewriting.html',
+          {inlineCss: true, rewriteUrlsInTemplates: true});
+      const style = dom5.query(
+          doc, matchers.styleMatcher, dom5.childNodesIncludeTemplate)!;
+      assert.isNotNull(style);
       assert.match(
-          dom5.getTextContent(styles[1]), /simple-style/, 'simple-style.css');
-      assert.match(
-          dom5.getTextContent(styles[2]),
-          /import-linked-style/,
-          'import-linked-style.css');
+          dom5.getTextContent(style),
+          /url\("style-rewriting\/styles\/unicorn.png"\)/);
     });
 
     test('Inlined styles have proper paths', async () => {